reorganize schema commands+utils
